refactor(api): delete FootbalistoClient, transforms.ts and clean up tests (#858)
Phase 4 of FootbalistoService migration: remove scaffolding now that all
handlers delegate to FootbalistoService.

- Delete FootbalistoClient tag, FootbalistoClientLive layer, and client.test.ts
- Delete transforms.ts and transforms.test.ts — functions moved into service.ts
- Remove FootbalistoMatch and FootbalistoMatchesArray from schemas.ts (unused)
- Extract PsdTeamClient for psd-sanity-sync (teams/members/staff fetch)
- Update index.ts to wire PsdTeamClientLive for sync cron handler

Closes #858
